overlord/devicestate: refactor and split into per-functionality files, drop dead code
As a prerequisite for the remodel work, attempt to refactor and split the
current code into files based on functionality. This should prevent
unctontrollable growth of devicestate and avoid the issues we have with
snapstate.
